requirements.txt
# 웹 프레임워크
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0

# 데이터베이스
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
pymysql==1.1.0
cryptography==41.0.4

# 인증
flask-jwt-extended==4.5.3

# API 및 스키마
marshmallow==3.20.1

# AWS
boto3==1.28.62

# HTTP 클라이언트
requests==2.31.0

# 환경 변수 및 유틸리티
python-dotenv==1.0.0
python-dateutil==2.8.2

# 서버 실행
gunicorn==21.2.0

# 로깅
colorlog==6.7.0